# docs/openapi.yaml
openapi: 3.0.3
info:
  title: CU Faculty Backend API
  version: "1.0.0"
  description: >
    REST API for the CU Faculty student hub. All endpoints under `/api` require a
    valid `bff.cookie` session cookie, which is validated against the Central
    University (CU) upstream on every request. The authenticated CU student id is
    used as the local user id. The only unauthenticated endpoint is `/health`.


    List endpoints are paginated and return a `{ data, total, limit, offset }`
    envelope. Pagination is controlled with the `limit` (default 20, max 100) and
    `offset` (default 0) query parameters.
security:
  - cookieAuth: []
...
paths:
...
  /api/me/edu-places:
    get:
      tags: [Me]
      summary: List my education places
      operationId: getMyEduPlaces
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Paged education places
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EduPlacePage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalError" }
    post:
      tags: [Me]
      summary: Create an education place
      operationId: createEduPlace
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EduPlaceRequest" }
      responses:
        "201":
          description: Created education place
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EduPlace" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/me/edu-places/{eduId}:
    put:
      tags: [Me]
      summary: Update an education place
      operationId: updateEduPlace
      parameters:
        - $ref: "#/components/parameters/EduId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EduPlaceRequest" }
      responses:
        "200":
          description: Updated education place
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EduPlace" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
    delete:
      tags: [Me]
      summary: Delete an education place
      operationId: deleteEduPlace
      parameters:
        - $ref: "#/components/parameters/EduId"
      responses:
        "204": { description: Deleted }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
...
  /api/me/events:
    get:
      tags: [Me]
      summary: List events I authored
      operationId: getMyEvents
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Paged events
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EventPage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalError" }
...
  /api/events:
    get:
      tags: [Events]
      summary: List events
      operationId: listEvents
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Paged events
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EventPage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalError" }
    post:
      tags: [Events]
      summary: Create an event
      description: >-
        Accepts multipart/form-data: a JSON `data` part with the event fields and
        an optional `photo` image part uploaded in the same request.
      operationId: createEvent
      requestBody:
        required: true
        description: >
          Actually sent as `multipart/form-data`: a required JSON `data` part
          matching the schema below, plus an optional `photo` image part
          (Content-Type must start with `image/`). Documented as `application/json`
          so the `data` schema renders; only the `data` part is shown here.
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateEventData" }
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Event" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/events/{id}:
    get:
      tags: [Events]
      summary: Get an event
      operationId: getEvent
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Event" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
    put:
      tags: [Events]
      summary: Update an event (author only)
      description: >-
        Accepts multipart/form-data with an optional JSON `data` part (event
        fields, partial update) and an optional `photo` image part. Supplying
        `photo` replaces the current one. Either or both parts may be omitted.
      operationId: updateEvent
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      requestBody:
        required: false
        description: >
          Actually sent as `multipart/form-data`: an optional JSON `data` part
          matching the schema below (partial update), plus an optional `photo`
          image part (Content-Type must start with `image/`) that replaces the
          current one. Documented as `application/json` so the `data` schema
          renders; only the `data` part is shown here.
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateEventData" }
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Event" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
    delete:
      tags: [Events]
      summary: Delete an event (author only)
      operationId: deleteEvent
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "204": { description: Deleted }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/events/{id}/responses:
    get:
      tags: [Events]
      summary: List responders to an event
      operationId: getEventResponders
      parameters:
        - $ref: "#/components/parameters/ResourceId"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Paged responders
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserPage" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalError" }
    post:
      tags: [Events]
      summary: Respond to an event
      description: Idempotent. Fails if the event is missing or a draft.
      operationId: respondToEvent
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "201":
          description: Event responded to
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Event" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
    delete:
      tags: [Events]
      summary: Remove my response to an event
      operationId: deleteMyEventResponse
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "204": { description: Removed }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: bff.cookie
      description: CU session cookie, validated against the CU upstream on every request.
  parameters:
    Limit:
      name: limit
      in: query
      required: false
      description: Page size (default 20, clamped to a maximum of 100).
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Offset:
      name: offset
      in: query
      required: false
      description: Number of items to skip (default 0).
      schema: { type: integer, minimum: 0, default: 0 }
...
    ResourceId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    EduId:
      name: eduId
      in: path
      required: true
      schema: { type: integer }
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing or rejected session cookie
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      description: >
        Normalized error envelope returned by every error response. "error" is a
        human-readable message (not for programmatic use); "code" is a stable,
        machine-readable identifier clients may branch on.
      properties:
        error: { type: string, example: "announcement not found" }
        code:
          type: string
          example: not_found
          enum:
            - bad_request
            - validation_error
            - unauthorized
            - forbidden
            - not_found
            - conflict
            - already_exists
            - invalid_reference
            - payload_too_large
            - upstream_error
            - unavailable
            - internal_error
      required: [error, code]
    PageMeta:
      type: object
      properties:
        total: { type: integer, description: Total number of matching items }
        limit: { type: integer }
        offset: { type: integer }
      required: [total, limit, offset]
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        photoUrl:
          type: string
          nullable: true
          description: Presigned, time-limited download URL for the user's photo.
        firstName: { type: string, maxLength: 31 }
        lastName: { type: string, maxLength: 31 }
        bio: { type: string, nullable: true, maxLength: 255 }
        birthdate: { type: string, format: date, example: "2000-01-31" }
        speciality: { type: string, nullable: true, maxLength: 63 }
        status: { type: string, nullable: true, maxLength: 127, description: Resolved status content }
        role: { type: string, enum: [user, admin] }
      required: [id, firstName, lastName, birthdate, role]
    EduPlace:
      type: object
      properties:
        id: { type: integer }
        universityId: { type: integer }
        universityName: { type: string, maxLength: 127 }
        grade: { $ref: "#/components/schemas/EduGrade" }
        level: { type: string, nullable: true, maxLength: 31 }
        specialization: { type: string, maxLength: 63 }
        startYear: { type: integer, minimum: -32768, maximum: 32767 }
        endYear: { type: integer, nullable: true, minimum: -32768, maximum: 32767 }
        isStudyingNow: { type: boolean }
      required: [id, universityId, universityName, grade, specialization, startYear, isStudyingNow]
    EduPlaceRequest:
      type: object
      properties:
        universityId: { type: integer }
        grade: { $ref: "#/components/schemas/EduGrade" }
        level: { type: string, nullable: true, maxLength: 31 }
        specialization: { type: string, maxLength: 63 }
        startYear: { type: integer, minimum: -32768, maximum: 32767 }
        endYear: { type: integer, nullable: true, minimum: -32768, maximum: 32767 }
        isStudyingNow: { type: boolean }
      required: [universityId, grade, specialization, startYear, isStudyingNow]
...
    CreateEventData:
      type: object
      required: [title, content, place, category, startsAt]
      properties:
        title: { type: string, maxLength: 127 }
        content: { type: string, maxLength: 255 }
        place: { type: string, maxLength: 63 }
        category: { $ref: "#/components/schemas/EventCategory" }
        startsAt: { type: string, format: date-time }
        registrationLink: { type: string, nullable: true, maxLength: 255 }
        isDraft: { type: boolean }
    UpdateEventData:
      type: object
      description: Partial update; omitted fields keep their current value
      properties:
        title: { type: string, maxLength: 127 }
        content: { type: string, maxLength: 255 }
        place: { type: string, maxLength: 63 }
        category: { $ref: "#/components/schemas/EventCategory" }
        startsAt: { type: string, format: date-time }
        registrationLink: { type: string, nullable: true, maxLength: 255 }
        isDraft: { type: boolean }
...
    Event:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string, maxLength: 127 }
        content: { type: string, maxLength: 255 }
        photoUrl: { type: string, nullable: true, description: Presigned download URL }
        place: { type: string, maxLength: 63 }
        category: { $ref: "#/components/schemas/EventCategory" }
        startsAt: { type: string, format: date-time }
        registrationLink: { type: string, nullable: true, maxLength: 255 }
        isDraft: { type: boolean }
        createdAt: { type: string, format: date-time }
        author: { $ref: "#/components/schemas/User" }
      required: [id, title, content, place, category, startsAt, isDraft, createdAt, author]
...
    EduGrade:
      type: string
      enum: [bachelor, master, specialist, postgraduate]
...
    EventCategory:
      type: string
      enum: [networking, professional, partner]
    UserPage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          properties:
            data: { type: array, items: { $ref: "#/components/schemas/User" } }
          required: [data]
    EduPlacePage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          properties:
            data: { type: array, items: { $ref: "#/components/schemas/EduPlace" } }
          required: [data]
...
    EventPage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          properties:
            data: { type: array, items: { $ref: "#/components/schemas/Event" } }
          required: [data]
...